Added class User and their Member variables

+++ RentCars/RentCars/Model/Car.cs
@@ -2,7 +2,7 @@
 {
     public class Car
     {
-        public int CID { get; set; }
+        public long CID { get; set; }
         public string CarName { get; set; }
         public string CarBrand { get; set; }
         public int Seats { get; set; }
